backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sse-starlette==1.6.5
pydantic==2.3.0
python-dotenv==1.0.0
groq==0.4.1
httpx==0.24.1
asyncio==3.4.3 